DCC Character Sheet
Setup the DCC character sheet with all functionality for generating a new character -- located within the 'dcc_character.py' file.

The 'dcc_gen.py' file uses tkinter as an example for how it can be used to generate a party of characters.

Note: If the character sheet file is missing, then a new one will be generated.